# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workout API",
    "version": "1.0.0",
    "description": "API for managing workouts with authentication"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    },
    {
      "url": "https://workoutapi-jl18.onrender.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "email": "test@test.com",
                "password": "123456"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "User already exists or invalid input"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "email": "test@test.com",
                "password": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "User not found"
          },
          "401": {
            "description": "Invalid password"
          }
        }
      }
    },
...
  }
}